requirements.txt
fastapi
uvicorn
python-dotenv
httpx
dropbox
PyPDF2
google-generativeai
streamlit
python-multipart
python-dotenv
